--- a/src/riblt/mapping.rs
+++ b/src/riblt/mapping.rs
@@ -8,22 +8,22 @@ use super::HashType;
 #[derive(Debug, Copy, Clone)]
 pub struct RandomMapping {
     /// PRNG state
-    prng: HashType,
+    prng: u8,
     /// The last index the symbol was mapped to
-    pub last_index: u64,
+    pub last_index: u16,
 }
 
 impl RandomMapping {
     pub fn new(prng: HashType) -> Self {
-        Self { prng, last_index: 0 }
+        Self { prng: prng as u8, last_index: 0 }
     }
 
     /// nextIndex returns the next index in the sequence.
-    pub fn next_index(&mut self) -> u64 {
+    pub fn next_index(&mut self) -> u16 {
         // Update the PRNG. TODO: prove that the following update rule gives us
         // high quality randomness, assuming the multiplier is coprime to 2^64.
-        let r = (self.prng as u64).wrapping_mul(0xda942042e4dd58b5);
-        self.prng = r as HashType;
+        let r = (self.prng as u16).wrapping_mul(0x58b5);
+        self.prng = r as u8;
         // Calculate the difference from the current index (s.lastIdx) to the
         // next index. See the paper for details. We use the approximated form
         //   diff = (1.5+i)((1-u)^(-1/2)-1)
@@ -32,8 +32,9 @@ impl RandomMapping {
         // that our u actually comes from sampling a random uint64 r, and then
         // dividing it by maxUint64, i.e., 1<<64. So we can replace (1-u)^
         // (-1/2) with 1<<32 / sqrt(r).
-        self.last_index += ((self.last_index as f64 + 1.5) *
-            ((1u64 << 32) as f64 / ((r+1) as f64).sqrt() - 1.0)).ceil() as u64;
+        let addend = ((self.last_index as f32 + 1.5) *
+            ((1u16 << 8) as f32 / ((r+1) as f32).sqrt() - 1.0)).ceil() as u16;
+        self.last_index = self.last_index.checked_add(addend).unwrap_or(u16::MAX);
         self.last_index
     }
 }