--- a/src/riblt/mod.rs
+++ b/src/riblt/mod.rs
@@ -1,7 +1,6 @@
 mod symbol;
 mod mapping;
 mod decoder;
-mod encoder;
 mod sketch;
 
 type HashType = u32;