--- a/src/riblt/sketch.rs
+++ b/src/riblt/sketch.rs
@@ -5,7 +5,7 @@
 //! sum quACK implementation in Go is available here:
 //! https://github.com/ygina/subset-reconciliation/.
 use super::HashType;
-use super::symbol::CodedSymbol;
+use super::symbol::{CodedSymbol, ADD, REMOVE};
 use super::mapping::RandomMapping;
 use super::decoder::Decoder;
 
@@ -48,8 +48,7 @@ impl Quack for IBLTQuackU32 {
         let mut m = RandomMapping::new(t);
         while (m.last_index as usize) < self.sketch.len() {
             let idx = m.last_index as usize;
-            self.sketch[idx].count += 1;
-            self.sketch[idx].hash ^= t;
+            self.sketch[idx].apply(t, ADD);
             m.next_index();
         }
     }
@@ -61,8 +60,7 @@ impl Quack for IBLTQuackU32 {
         let mut m = RandomMapping::new(t);
         while (m.last_index as usize) < self.sketch.len() {
             let idx = m.last_index as usize;
-            self.sketch[idx].count -= 1;
-            self.sketch[idx].hash ^= t;
+            self.sketch[idx].apply(t, REMOVE);
             m.next_index();
         }
     }
@@ -78,8 +76,8 @@ impl Quack for IBLTQuackU32 {
         self.count = self.count.wrapping_sub(s2.count);
 
         for i in 0..self.sketch.len() {
-            self.sketch[i].count -= s2.sketch[i].count;
-            self.sketch[i].hash ^= s2.sketch[i].hash;
+            let x = s2.sketch[i];
+            self.sketch[i].apply(x.hash, x.count.wrapping_neg());
         }
     }
 
@@ -112,7 +110,7 @@ mod test{
 
     #[test]
     fn test_fixed_encode_and_decode() {
-        let sizes = vec![10, 20, 40, 100, 1000, 10000, 50000, 100000];
+        let sizes = vec![10, 20, 40, 100, 200, (u8::MAX - 1).into()];
         for size in sizes {
             let nlocal = size;
             let ncommon = size;