In terms of number of bytes, encode time, and decode time. Involves using smaller bit-width values in the random mapping generator, the counts, and removing unnecessary decode operations for subset-only reconciliation.

Since a single quACK needs to fit in a packet MTU, for a 1500-byte MTU and 42-byte UDP header payload offset, we have a 1458-byte UDP payload. There are 4 bytes for the packet count, and 4 bytes for the last inserted value. Each coded symbol is 5 bytes (4-byte identifier and 1-byte count), leaving generously at most `(1458-8)/5 = 290` source symbols. The quACK requires at least one source symbol per error to decode---this is a theoretical bound---so the 290 source symbols is also an upper bound on the number of errors. Hence, we use 1 byte for the coded symbol count which is slightly less than 290.

IBLT results before:

```
$ ./target/release/examples/benchmark_riblt iblt --encode
BenchmarkIBLTEncode/m=10         2097152           77ns/op
BenchmarkIBLTEncode/m=20         1048576           97ns/op
BenchmarkIBLTEncode/m=40         1048576          118ns/op
BenchmarkIBLTEncode/m=80         1048576          137ns/op
BenchmarkIBLTEncode/m=160        1048576          157ns/op
BenchmarkIBLTEncode/m=200        1048576          163ns/op
BenchmarkIBLTEncode/m=240        1048576          168ns/op
BenchmarkIBLTEncode/m=280        1048576          173ns/op
```

```
$ ./target/release/examples/benchmark_riblt iblt --decode
BenchmarkIBLTDecode/d=1           524288          206ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkIBLTDecode/d=2           262144          419ns/op     0.686 succ@2x     0.937 succ@4x     2.128 symbols/diff
BenchmarkIBLTDecode/d=4           131072          851ns/op     0.548 succ@2x     0.914 succ@4x     2.367 symbols/diff
BenchmarkIBLTDecode/d=8            65536         1829ns/op     0.580 succ@2x     0.975 succ@4x     2.137 symbols/diff
BenchmarkIBLTDecode/d=10           65536         2433ns/op     0.653 succ@2x     0.986 succ@4x     1.999 symbols/diff
BenchmarkIBLTDecode/d=20           32768         4537ns/op     0.889 succ@2x     0.999 succ@4x     1.665 symbols/diff
BenchmarkIBLTDecode/d=40           16384         9204ns/op     0.980 succ@2x     1.000 succ@4x     1.538 symbols/diff
BenchmarkIBLTDecode/d=80            8192        18658ns/op     0.999 succ@2x     1.000 succ@4x     1.475 symbols/diff
BenchmarkIBLTDecode/d=160           4096        41656ns/op     1.000 succ@2x     1.000 succ@4x     1.428 symbols/diff
BenchmarkIBLTDecode/d=255           2048        70639ns/op     1.000 succ@2x     1.000 succ@4x     1.407 symbols/diff
```

IBLT results after:

```
$ ./target/release/examples/benchmark_riblt iblt --encode
BenchmarkIBLTEncode/m=10         4194304           45ns/op          71 bytes
BenchmarkIBLTEncode/m=20         2097152           62ns/op         121 bytes
BenchmarkIBLTEncode/m=40         2097152           81ns/op         221 bytes
BenchmarkIBLTEncode/m=80         1048576          101ns/op         421 bytes
BenchmarkIBLTEncode/m=160        1048576          121ns/op         821 bytes
BenchmarkIBLTEncode/m=200        1048576          127ns/op        1021 bytes
BenchmarkIBLTEncode/m=240        1048576          131ns/op        1221 bytes
BenchmarkIBLTEncode/m=280        1048576          134ns/op        1421 bytes
```

```
$ ./target/release/examples/benchmark_riblt iblt --decode
BenchmarkIBLTDecode/d=1          1048576          122ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkIBLTDecode/d=2          1048576          175ns/op     0.844 succ@2x     0.984 succ@4x     1.414 symbols/diff
BenchmarkIBLTDecode/d=4           262144          405ns/op     0.656 succ@2x     0.938 succ@4x     2.172 symbols/diff
BenchmarkIBLTDecode/d=8           131072          828ns/op     0.812 succ@2x     1.000 succ@4x     1.852 symbols/diff
BenchmarkIBLTDecode/d=16           65536         1681ns/op     1.000 succ@2x     1.000 succ@4x     1.594 symbols/diff
BenchmarkIBLTDecode/d=20           65536         2449ns/op     0.969 succ@2x     1.000 succ@4x     1.595 symbols/diff
BenchmarkIBLTDecode/d=40           32768         5107ns/op     1.000 succ@2x     1.000 succ@4x     1.581 symbols/diff
BenchmarkIBLTDecode/d=80           16384        11015ns/op     1.000 succ@2x     1.000 succ@4x     1.472 symbols/diff
BenchmarkIBLTDecode/d=160           4096        24478ns/op     1.000 succ@2x     1.000 succ@4x     1.427 symbols/diff
BenchmarkIBLTDecode/d=255           4096        42956ns/op     1.000 succ@2x     1.000 succ@4x     1.459 symbols/diff
```

I think the IBLT decode results might be skewed by worst case since it is an average and not a median.

Power sum results (unchanged):

```
$ ./target/release/examples/benchmark_riblt power-sum --encode
BenchmarkPowerSumEncode/m=10         4194304           31ns/op          61 bytes
BenchmarkPowerSumEncode/m=20         2097152           66ns/op         101 bytes
BenchmarkPowerSumEncode/m=40         1048576          139ns/op         181 bytes
BenchmarkPowerSumEncode/m=80          524288          283ns/op         341 bytes
BenchmarkPowerSumEncode/m=160         262144          573ns/op         661 bytes
BenchmarkPowerSumEncode/m=200         262144          717ns/op         821 bytes
BenchmarkPowerSumEncode/m=240         131072          862ns/op         981 bytes
BenchmarkPowerSumEncode/m=280         131072         1006ns/op        1141 bytes
```

```
$ ./target/release/examples/benchmark_riblt power-sum --decode
BenchmarkPowerSumDecode/d=1          1048576          164ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=2           262144          496ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=4            65536         2190ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=8            16384        11397ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=16            2048        60904ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=20            1024       100264ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=40             256       446926ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=80              64      1846770ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=160             16      7503647ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
BenchmarkPowerSumDecode/d=255              8     19168118ns/op     1.000 succ@2x     1.000 succ@4x     1.000 symbols/diff
```

IBLT encoding and decoding both scale way better, though we typically don't have such large `m` and `d`. IBLT decoding is always faster, granted we have enough source symbols. IBLT encoding is faster for `m>20`, though note we need more source symbols than the number of errors to decode with high probability. In actual decoding time, in IBLT we need to map  the decoded identifiers back to the original packet, whereas in power-sum we are iterating through the log. The actual decoding time also includes encoding every packet in addition to the subset reconciliation step measured here. There are definitely scenarios in which the IBLT could be a faster choice (e.g., enough memory at the proxy to encode larger values for the worst case, frequent decodes which are more efficient, not limited by link overhead so can send frequent quACKs).